Added a feature that lets users choose their favorite color during onboarding, because we all know they'll never pick green.
(auto-msg qwen2.5-coder:7b)
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   test.lua

Implemented a feature that counts the number of invisible unicorns in the database

(auto-msg qwen2.5-coder:7b)
On branch master
Your branch is ahead of 'origin/master' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   test.lua